+++ plugin/src/main/kotlin/dev/inmo/plagubot/Plugin.kt
@@ -5,6 +5,7 @@ import dev.inmo.tgbotapi.extensions.behaviour_builder.BehaviourContext
 import dev.inmo.tgbotapi.types.BotCommand
 import dev.inmo.tgbotapi.updateshandlers.FlowsUpdatesFilter
 import kotlinx.coroutines.CoroutineScope
+import kotlinx.serialization.Serializable
 import org.jetbrains.exposed.sql.Database
 
 /**
@@ -12,6 +13,7 @@ import org.jetbrains.exposed.sql.Database
  * to set up short name for your plugin. Besides, simple name of your class will be used as key for deserialization
  * too.
  */
+@Serializable(PluginSerializer::class)
 interface Plugin {
     /**
      * In case you want to publish some processed by your plugin commands, you can provide it via this method