+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.1.3
+
+* `Versions`
+  * `tgbotapi`: `0.32.6` -> `0.32.7`
+* `Bot`
+  * `initPlaguBot` now will return `Job`
+* `Plugin`
+  * Plugin serializer
+
 ## 0.1.2
 
 * `Versions`